backend/requirements.txt
flask
flask-cors
requests
python-dotenv
exa-py
